+++ FamilyTreeFIR/FamilyTree/家谱服务/我的互助/MyHelpViewController.m
@@ -95,12 +95,12 @@ typedef enum : NSUInteger {
     NSInteger index = seg.selectedSegmentIndex;
     switch (index) {
         case 0:
-            MYLog(@"%ld",index);
+            MYLog(@"%ld",(long)index);
             self.isMyPost = YES;
             [self.infoTB reloadData];
             break;
         case 1:
-            MYLog(@"%ld",index);
+            MYLog(@"%ld",(long)index);
             self.isMyPost = NO;
             if (self.myAttentionArr.count == 0) {
                 [self getMYAttentionData];
