+++ FamilyTreeFIR/FamilyTree/家谱服务/祭祀/墓园/建园/CreateCemViewController.m
@@ -40,7 +40,7 @@
 
 - (void)viewDidLoad {
     [super viewDidLoad];
-    MYLog(@"%ld",self.CeId);
+    MYLog(@"%ld",(long)self.CeId);
     [self initData];
     [self initUI];
 }
